+++ src/frontend/js/utils/mptt/index.spec.ts
@@ -1,4 +1,4 @@
-import { isMPTTParentOf } from '.';
+import { getMPTTChildrenPathMatcher, isMPTTParentOf } from '.';
 
 describe('utils/mptt', () => {
   // NB: We're not testing isMPTTChildOf because it is logically the exact same function as `isMPTTParentOf`,
@@ -22,6 +22,23 @@ describe('utils/mptt', () => {
       expect(isMPTTParentOf('P-00020004', 'L-000200040003')).toEqual(true);
       expect(isMPTTParentOf('P-00020004', 'L-000200040003000C0009')).toEqual(
         true,
+      );
+    });
+  });
+
+  describe('getMPTTChildrenPathMatcher()', () => {
+    it('throws when the passed key is not an MPTT path', () => {
+      expect(() => getMPTTChildrenPathMatcher('en')).toThrow();
+      expect(() => getMPTTChildrenPathMatcher('L-000')).toThrow();
+    });
+
+    it('returns a regex-string that will match paths for children of the passed MPTT path entity', () => {
+      expect(getMPTTChildrenPathMatcher('P-0001')).toEqual('.-0001.{4,}');
+      expect(getMPTTChildrenPathMatcher('L-0001000C')).toEqual(
+        '.-0001000C.{4,}',
+      );
+      expect(getMPTTChildrenPathMatcher('P-000100090006000C')).toEqual(
+        '.-000100090006000C.{4,}',
       );
     });
   });